backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
huggingface_hub
langchain-core
langgraph
redis
